+++ RenderApp/Geometry/Displacement/SineDisplacement.cpp
@@ -11,18 +11,18 @@ SineDisplacement::~SineDisplacement() { };
 /// Returns a displacement in [-scale, scale]
 float SineDisplacement::getDisplacement(float u, float v) {
 	if (uTimes > 0 && vTimes > 0)
-		return (sin(u * uTimes * 2 * PI) + sin(v * vTimes * 2 * PI)) * scale;
+		return static_cast<float>(sin(u * uTimes * 2 * PI) + sin(v * vTimes * 2 * PI)) * scale;
 	if (uTimes > 0)
-		return sin(u * uTimes * 2 * PI) * scale;
+		return static_cast<float>(sin(u * uTimes * 2 * PI)) * scale;
 	if (vTimes > 0)
-		return sin(v * vTimes * 2 * PI) * scale;
+		return static_cast<float>(sin(v * vTimes * 2 * PI)) * scale;
 	return 0; // Should never occurr
 }
 
 float SineDisplacement::getDerivativeU(float u, float v) {
-	return -cos(u * uTimes * 2 * PI) * scale;
+	return static_cast<float>(-cos(u * uTimes * 2 * PI)) * scale;
 }
 
 float SineDisplacement::getDerivativeV(float u, float v) {
-	return -cos(v * vTimes * 2 * PI) * scale;
+	return static_cast<float>(-cos(v * vTimes * 2 * PI)) * scale;
 }
