+++ SamplingApp/Sampler/Sampler.h
@@ -7,6 +7,6 @@
 
 class Sampler {
 	public:
-		virtual void generateSamples(int nSamplesSqrt, std::vector<Math::Vec2f>& sampleStorage) = 0;
+		virtual void generateSamples(size_t nSamplesSqrt, std::vector<Math::Vec2f>& sampleStorage) = 0;
 		virtual const char* getName() = 0;
 };
