+++ SamplingApp/Sampler/UniformSampler.h
@@ -6,6 +6,6 @@
 #include <vector>
 
 class UniformSampler : public Sampler {
-	virtual void generateSamples(int nSamplesSqrt, std::vector<Math::Vec2f>& sampleStorage) override;
+	virtual void generateSamples(size_t nSamplesSqrt, std::vector<Math::Vec2f>& sampleStorage) override;
 	virtual const char* getName() override;
 };
